pyproject.toml
[tool.poetry]
name = "music-platform"
version = "0.1.0"
description = "Music Platform"
authors = ["7oSkaaa <ahmed.7oskaa@gmail.com>"]
readme = "README.md"
packages = [{include = "music_platform"}]

[tool.poetry.dependencies]
python = "^3.10"
dotenv = "^0.0.5"
django-imagekit = "^4.1.0"
pillow = "^9.3.0"
django = "^4.1.3"
pilkit = "^2.0"
django-model-utils = "^4.2.0"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.1"
knox = "^0.1.14"
django-rest-knox = "^4.2.0"
pyyaml = "^6.0"
pytest-django = "^4.5.2"
django-dotenv = "^1.4.2"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
